requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
twilio==9.2.3
requests==2.32.3
openai==1.51.2
python-multipart==0.0.9
redis==5.0.7
tenacity==8.5.0
httpx==0.27.2
